==> repeat_with_no_schizophrenia/02_milo_oligos.Rmd <==
---
title: "Differential abundance testing with Milo - Mouse gastrulation example"
author:
  - Emma Dann
  - Mike Morgan
  - Modified by Nadine bestardCuche
output:
  BiocStyle::html_document:
    toc_float: true
  BiocStyle::pdf_document: default
package: miloR
vignette: |
  %\VignetteIndexEntry{Differential abundance testing with Milo - Mouse gastrulation example}
  %\VignetteEngine{knitr::rmarkdown}
  %\VignetteEncoding{UTF-8}
---

```{r, include = FALSE}
knitr::opts_chunk$set(
  collapse = FALSE,
  message=FALSE
)
```

```{r include=FALSE}
.libPaths("U:/Datastore/CMVM/scs/groups/Williamsdata/Sunniva/Nadine/renv/library/R-4.4/x86_64-w64-mingw32")
```

```{r setup, message=FALSE, warning=FALSE}
library(miloR)
library(SingleCellExperiment)
library(scater)
library(scran)
library(dplyr)
library(patchwork)
library(here)
library(Seurat)
library(tidyseurat)
```


# Load data

```{r}
oligo <- readRDS(here("data/repeat_with_no_schizophrenia/oligocytes.RDS"))
oligo_sce <- as.SingleCellExperiment(oligo, assay = "integrated")
```


# Differential abundance testing

## Create a Milo object

For differential abundance analysis on graph neighbourhoods we first construct a `Milo` object. This extends the [`SingleCellExperiment`](https://bioconductor.org/packages/release/bioc/html/SingleCellExperiment.html) class to store information about neighbourhoods on the KNN graph. 

```{r}
oligo_milo <- Milo(oligo_sce)
oligo_milo
```


## Construct KNN graph

We need to add the KNN graph to the Milo object. This is stored in the `graph` slot, in [`igraph`](https://igraph.org/r/) format. The `miloR` package includes functionality to build and store the graph from the PCA dimensions stored in the `reducedDim` slot. In this case, we specify that we want to build the graph from the MNN corrected PCA dimensions.

For graph building you need to define a few parameters:

- `d`: the number of reduced dimensions to use for KNN refinement. We recommend using the same $d$ used for KNN graph building, or to select PCs by inspecting the [scree plot](http://bioconductor.org/books/release/OSCA/dimensionality-reduction.html#choosing-the-number-of-pcs).
- `k`: this  affects the power of DA testing, since we need to have enough cells from each sample represented in a neighbourhood to estimate the variance between replicates. On the other side, increasing $k$ too much might lead to over-smoothing. We suggest to start by using the same value for $k$ used for KNN graph building for clustering and UMAP visualization. We will later use some heuristics to evaluate whether the value of $k$ should be increased.

```{r}
oligo_milo <- buildGraph(oligo_milo, k = 30, d = 13, reduced.dim = "PCA")
```

Alternatively, one can add a precomputed KNN graph (for example constructed with Seurat or scanpy) to the `graph` slot using the adjacency matrix, through the helper function `buildFromAdjacency`.

<!-- Alternatively, if you already have a KNN graph (for example constructed with Seurat/scanpy) you can add it from the adjacency matrix. -->

<!-- ```{r} -->
<!-- # ## Build up a mock SNN graph made with Seurat -->
<!-- # pca_df <- reducedDim(traj_milo, "PCA") -->
<!-- # rownames(pca_df) <- traj_milo$cell_id -->
<!-- # snn_graph <- FindNeighbors(pca_df)[["snn"]] -->
<!-- #  -->
<!-- # graph(traj_milo) <-  graph(buildFromAdjacency(snn_graph, k=10)) -->
<!-- ``` -->

## Defining representative neighbourhoods on the KNN graph

We define the neighbourhood of a cell, the index, as the group of cells connected by an edge in the KNN graph to the index cell. For efficiency, we don't test for DA in the neighbourhood of every cell, but we sample as indices a subset of representative cells, using a KNN sampling algorithm used by [Gut et al. 2015](https://www.nature.com/articles/nmeth.3545). 

As well as $d$ and $k$, for sampling we need to define a few additional parameters:

- `prop`: the proportion of cells to randomly sample to start with. We suggest using `prop=0.1` for datasets of less than 30k cells. For bigger datasets using `prop=0.05` should be sufficient (and makes computation faster).
- `refined`: indicates whether you want to use the sampling refinement algorithm, or just pick cells at random. The default and recommended way to go is to use refinement. The only situation in which you might consider using `random` instead, is if you have batch corrected your data with a graph based correction algorithm, such as [BBKNN](https://github.com/Teichlab/bbknn), but the results of DA testing will be suboptimal.

```{r}
oligo_milo <- makeNhoods(oligo_milo, prop = 0.1, k = 10, d=13, refined = TRUE, reduced_dims = "PCA", refinement_scheme="graph") 
```

Once we have defined neighbourhoods, we plot the distribution of neighbourhood sizes (i.e. how many cells form each neighbourhood) to evaluate whether the value of $k$ used for graph building was appropriate. We can check this out using the `plotNhoodSizeHist` function. 

As a rule of thumb we want to have an average neighbourhood size over 5 x N_samples. If the mean is lower, or if the distribution is 

```{r}
plotNhoodSizeHist(oligo_milo)
```

## Counting cells in neighbourhoods

_Milo_ leverages the variation in cell numbers between replicates for the same experimental condition to test for differential abundance. Therefore we have to count how many cells from each sample are in each neighbourhood. We need to use the cell metadata and specify which column contains the sample information.

```{r}
oligo_milo <- countCells(oligo_milo, meta.data = as.data.frame(colData(oligo_milo)), sample="ProcessNumber")
```

This adds to the `Milo` object a $n \times m$ matrix, where $n$ is the number of neighbourhoods and $m$ is the number of experimental samples. Values indicate the number of cells from each sample counted in a neighbourhood. This count matrix will be used for DA testing.

```{r}
head(nhoodCounts(oligo_milo))
```

## Defining experimental design

Now we are all set to test for differential abundance in neighbourhoods. We implement this hypothesis testing in a generalized linear model (GLM) framework, specifically using the Negative Binomial GLM implementation in [`edgeR`](https://bioconductor.org/packages/release/bioc/html/edgeR.html).

We first need to think about our experimental design. The design matrix should match each sample to the experimental condition of interest for DA testing. In this case, we want to detect DA between oligonic stages, stored in the `stage` column of the dataset `colData`. We also include the `sequencing.batch` column in the design matrix. This represents a known technical covariate that we want to account for in DA testing. 

```{r}
oligo_design <- data.frame(colData(oligo_milo))[,c("ProcessNumber", "donor_status", "Tissue")]

## Convert info from integer to factor
oligo_design$Tissue <- as.factor(oligo_design$Tissue) 
oligo_design$donor_status <- as.factor(oligo_design$donor_status)
oligo_design$donor_status <- relevel(oligo_design$donor_status, "Ctrl")
# simplify 
oligo_design <- distinct(oligo_design)
rownames(oligo_design) <- oligo_design$ProcessNumber

#oligo_design
```


<!-- ## Computing neighbourhood connectivity -->
<!-- not needed as per github issue as(<dgTMatrix>, "dgCMatrix") is deprecated  -->
<!-- Milo uses an adaptation of the Spatial FDR correction introduced by [cydar](https://bioconductor.org/packages/release/bioc/html/cydar.html), where we correct p-values accounting for the amount of overlap between neighbourhoods. Specifically, each hypothesis test P-value is weighted by the reciprocal of the kth nearest neighbour distance. To use this statistic we first need to store the distances between nearest neighbors in the Milo object. This is done by the `calcNhoodDistance` function -->
<!-- (N.B. this step is the most time consuming of the analysis workflow and might take a couple of minutes for large datasets). -->

<!-- ```{r} -->
<!-- oligo_milo <- calcNhoodDistance(oligo_milo, d=13, reduced.dim = "PCA") -->
<!-- ``` -->

## Testing

Now we can do the DA test, explicitly defining our experimental design. In this case, we want to test for differences between experimental stages, while accounting for the variability between technical batches (You can find more info on how to use formulas to define a testing design in R [here](https://r4ds.had.co.nz/model-basics.html#formulas-and-model-families))

```{r}
da_results <- testNhoods(oligo_milo,  design = ~ Tissue + donor_status, design.df = oligo_design,  fdr.weighting="graph-overlap")
head(da_results)
```

This calculates a Fold-change and corrected P-value for each neighbourhood, which indicates whether there is significant differential abundance between developmental stages. The main statistics we consider here are: 

- `logFC`: indicates the log-Fold change in cell numbers between samples from E7.5 and samples from E7.0
- `PValue`: reports P-values before FDR correction
- `SpatialFDR`: reports P-values corrected for multiple testing accounting for overlap between neighbourhoods

```{r}
da_results %>%
  arrange(SpatialFDR) %>%
  head() 
```

# Inspecting DA testing results

We can start inspecting the results of our DA analysis from a couple of standard diagnostic plots. 
We first inspect the distribution of uncorrected P values, to verify that the test was balanced.

```{r}
ggplot(da_results, aes(PValue)) + geom_histogram(bins=50)
```

Then we visualize the test results with a volcano plot (remember that each point here represents a neighbourhood, _not_ a cell).

```{r, dev="jpeg"}
ggplot(da_results, aes(logFC, -log10(SpatialFDR))) + 
  geom_point() +
  geom_hline(yintercept = 1) ## Mark significance threshold (10% FDR)
```


To visualize DA results relating them to the embedding of single cells, we can build an abstracted graph of neighbourhoods that we can superimpose on the single-cell embedding. Here each node represents a neighbourhood, while edges indicate how many cells two neighbourhoods have in common. Here the layout of nodes is determined by the position of the index cell in the UMAP embedding of all single-cells. The neighbourhoods displaying significant DA are colored by their log-Fold Change.

```{r, fig.width=15, fig.height=8, dev="jpeg"}
oligo_milo <- buildNhoodGraph(oligo_milo)

## Plot single-cell UMAP
umap_pl <- plotReducedDim(oligo_milo, dimred = "UMAP", colour_by="donor_status", text_by = "ol_clusters_named_without_skf", 
                          text_size = 3, point_size=0.5) +
  guides(fill="none")

## Plot neighbourhood graph
nh_graph_pl <- plotNhoodGraphDA(oligo_milo, da_results, layout="UMAP",alpha=0.1) 
  
umap_pl + nh_graph_pl +
  plot_layout(guides="collect")
```

We might also be interested in visualizing whether DA is particularly evident in certain cell types. To do this, we assign a cell type label to each neighbourhood by finding the most abundant cell type within cells in each neighbourhood. We can label neighbourhoods in the results `data.frame` using the function `annotateNhoods`. This also saves the fraction of cells harbouring the label.

```{r}
da_results <- annotateNhoods(oligo_milo, da_results, coldata_col = "ol_clusters_named_without_skf")
head(da_results)
```

While neighbourhoods tend to be homogeneous, we can define a threshold for `celltype_fraction` to exclude neighbourhoods that are a mix of cell types. 

```{r}
ggplot(da_results, aes(ol_clusters_named_without_skf_fraction)) + geom_histogram(bins=50)
```
```{r}
da_results$ol_clusters_named_without_skf <- ifelse(da_results$ol_clusters_named_without_skf_fraction < 0.7, "Mixed", da_results$ol_clusters_named_without_skf)
```

Now we can visualize the distribution of DA Fold Changes in different cell types

```{r, fig.height=7, fig.width=7, dev="jpeg"}
da_results$ol_clusters_named_without_skf <- factor(da_results$ol_clusters_named_without_skf, levels= c("Mixed", "Oligo_Cedar","Oligo_Rowan","Oligo_Birch","Oligo_Maple", "Oligo_Elder", 
"Oligo_Oak", "OPC_Pine", "OPC_Fir", "COPs"))
plotDAbeeswarm(da_results, group.by = "ol_clusters_named_without_skf")
ggsave(here("outs/repeat_with_no_schizophrenia/abundance/milo_oligo.pdf"))
```
