==> repeat_with_no_schizophrenia/05_regions.Rmd <==
---
title: "05_regions"
author: "Nadine Bestard"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Libraries
```{r}
library(Seurat)
library(MAST)
library(ggsci)
library(here)
library(dplyr)
library(tibble)
library(tidyseurat)
```

#set colour palette
```{r}
mypal <- pal_npg("nrc", alpha = 0.7)(10)
mypal2 <-pal_tron("legacy", alpha = 0.7)(7)
mypal3 <- pal_lancet("lanonc", alpha = 0.7)(9)
mypal4 <- pal_simpsons(palette = c("springfield"), alpha = 0.7)(16)
mypal5 <- pal_rickandmorty(palette = c("schwifty"), alpha = 0.7)(6)
mypal6 <- pal_futurama(palette = c("planetexpress"), alpha = 0.7)(5)
mypal7 <- pal_startrek(palette = c("uniform"), alpha = 0.7)(5)
mycoloursP<- c(mypal, mypal2, mypal3, mypal4, mypal5, mypal6, mypal7)
#show_col(mycoloursP, labels =F)
```
# load objects
```{r}
mglia <- readRDS(here("data/repeat_with_no_schizophrenia/microglia.RDS"))
DefaultAssay(mglia) <- "RNA"
astro <-  readRDS(here("data/repeat_with_no_schizophrenia/astrocytes.RDS"))
DefaultAssay(astro) <- "RNA"
oligo <-  readRDS(here("data/repeat_with_no_schizophrenia/oligodendrocytes.RDS"))
DefaultAssay(oligo) <- "RNA"
```

# Compute DE global
```{r}
Idents(oligo) <- "donor_status"
oligo_de <- FindMarkers(oligo, ident.1 = "HD", logfc.threshold = 0.8, test.use = "MAST"  )
Idents(mglia) <- "donor_status"
mglia_de <- FindMarkers(mglia, ident.1 = "HD", logfc.threshold = 0.8, test.use = "MAST"  )
Idents(astro) <- "donor_status"
astro_de <- FindMarkers(astro, ident.1 = "HD", logfc.threshold = 0.8, test.use = "MAST"  )
```

## save
```{r}
de_list <- list(oligo = oligo_de, mglia = mglia_de, astro = astro_de)
de_list_filtered <- de_list
for(celltype in names(de_list)){
  de <- de_list[[celltype]]
de <- de %>%  
    filter(p_val_adj < 0.05) %>% 
    rownames_to_column("gene")
  write.csv(de, here("outs", "repeat_with_no_schizophrenia", "DE_HD_Cntrl", paste0( "HDvsCntrl_DE_in_", celltype,".csv")))
de_list_filtered[[celltype]] <- de
}


```

#  Compute DE per tissue 
```{r}
tissues <- unique(oligo$Tissue)
de_big <- data.frame()
list_srt <- list(oligo = oligo, mglia = mglia, astro = astro)
for(celltype in names(list_srt)){
  for(tissue in tissues){
    print(celltype)
    print(tissue)
    celltype_tissue <- list_srt[[celltype]]  %>% filter(Tissue == tissue)
    de <- FindMarkers(celltype_tissue, ident.1 = "HD", logfc.threshold = 0.8, test.use = "MAST"  )
    de <- de %>% mutate(tissue = tissue) %>% 
      mutate(celltype = celltype) %>% 
    filter(p_val_adj < 0.05) %>% 
      rownames_to_column("gene")
    write.csv(de, here("outs", paste0( "HDvsCntrl_DE_in_",celltype, tissue,".csv")))
    
    de_big <- rbind(de_big, de)
  }
}
write.csv(de_big, here("outs/repeat_with_no_schizophrenia/DE_HD_Cntrl/HDvsCntrl_DE_in_celltypes_tissues.csv"))
```



# Volcano Plots
i build a function to plot all the volcanos with the same parametres. 
8 top genes (FC top, no matter if down or up), axis adjusted with maximum value so all the volcanos have the same axis and it's easy to compare btw them
```{r}
library(EnhancedVolcano)
```

```{r}
plotVolcano <- function(df, colour){
top_genes <- arrange(df, desc(abs(avg_log2FC))) %>% dplyr::slice(1:8) %>% select(gene) %>% pull()
EnhancedVolcano(df,
                lab = df$gene,
                x = "avg_log2FC",
                y = "p_val_adj", 
                labSize = 3.5,
                FCcutoff = 0.8,
                pCutoff = 0.05,
                pointSize = 3,
                raster = F,
                #shapeCustom = keyvals.shape,
                #colCustom = df[,"colour"],
                selectLab = top_genes,
                boxedLabels = T,         
                drawConnectors = T,
                maxoverlapsConnectors = Inf,
                max.overlaps = Inf,
                labFace = "bold",
                title = NULL,
                subtitle =NULL,
                legendLabSize = 18,
                legendIconSize = 4,
                legendPosition = "right",
               # xlim = c((min(oligo_cb$avg_log2FC) - .05), (max(oligo_cb$avg_log2FC + 0.5)))
               xlim = c(-9, 9)
               ) +
  scale_colour_manual(values = colour ) +
  ylim(c(0,300)) +
   theme(legend.position="none")
}

```

Oligos
```{r}
library(EnhancedVolcano)
oligo_de <- de_big %>% 
  filter(celltype == "oligo") 
# separate tissues
# CB 22
oligo_cb <- oligo_de %>% filter(tissue == "CB") 
colour <-  mycoloursP[22]

p1 <- plotVolcano(oligo_cb, mycoloursP[22])
ggsave(here("outs/repeat_with_no_schizophrenia/regions/oligo_cb.pdf"), p1, width = 7, height = 4)

# HC 16
oligo_hc <- oligo_de %>% filter(tissue == "HC") 
colour <-  mycoloursP[16]

p1 <- plotVolcano(oligo_hc, colour)
ggsave(here("outs/repeat_with_no_schizophrenia/regions/oligo_hc.pdf"), p1, width = 6, height = 4)

# CN 18
oligo_cn <- oligo_de %>% filter(tissue == "CN") 
colour <-  mycoloursP[18]

p1 <- plotVolcano(oligo_cn, colour)
ggsave(here("outs/repeat_with_no_schizophrenia/regions/oligo_cn.pdf"), p1, width = 7, height = 4)

#FrCx 8
oligo_f <- oligo_de %>% filter(tissue == "FrCx") 
colour <-  mycoloursP[8]

p1 <- plotVolcano(oligo_f, colour)
ggsave(here("outs/repeat_with_no_schizophrenia/regions/oligo_frcx.pdf"), p1, width = 7, height = 4)
```
```{r include=FALSE}
# obtain values for DeepVen
mglia_de %>% filter(tissue == "FrCx") %>% .$gene %>% dput()
print("hey")
```

Microglia
```{r}
library(EnhancedVolcano)
mglia_de <- de_big %>% 
  filter(celltype == "mglia") 
# separate tissues
# CB 22
cb <- mglia_de %>% filter(tissue == "CB") 
colour <-  mycoloursP[22]

p1 <- plotVolcano(cb, mycoloursP[22])
ggsave(here("outs/repeat_with_no_schizophrenia/regions/mglia_cb.pdf"), p1, width = 7, height = 4)

# HC 16
hc <- mglia_de %>% filter(tissue == "HC") 
colour <-  mycoloursP[16]

p1 <- plotVolcano(hc, colour)
ggsave(here("outs/repeat_with_no_schizophrenia/regions/mglia_hc.pdf"), p1, width = 7, height = 4)

# CN 18
cn <- mglia_de %>% filter(tissue == "CN") 
colour <-  mycoloursP[18]

p1 <- plotVolcano(cn, colour)
ggsave(here("outs/repeat_with_no_schizophrenia/regions/mglia_cn.pdf"), p1, width = 7, height = 4)

#FrCx 8
fc <- mglia_de %>% filter(tissue == "FrCx") 
colour <-  mycoloursP[8]

p1 <- plotVolcano(fc, colour)
ggsave(here("outs/repeat_with_no_schizophrenia/regions/mglia_frcx.pdf"), p1, width = 7, height = 4)
```

Astroctyes
```{r}
library(EnhancedVolcano)
astro_de <- de_big %>% 
  filter(celltype == "astro") 
# separate tissues
# CB 22
cb <- astro_de %>% filter(tissue == "CB") 
colour <-  mycoloursP[22]

p1 <- plotVolcano(cb, mycoloursP[22])
ggsave(here("outs/repeat_with_no_schizophrenia/regions/astro_cb.pdf"), p1, width = 7, height = 4)

# HC 16
hc <- astro_de %>% filter(tissue == "HC") 
colour <-  mycoloursP[16]

p1 <- plotVolcano(hc, colour)
ggsave(here("outs/repeat_with_no_schizophrenia/regions/astro_hc.pdf"), p1, width = 7, height = 4)

# CN 18
cn <- astro_de %>% filter(tissue == "CN") 
colour <-  mycoloursP[18]

p1 <- plotVolcano(cn, colour)
ggsave(here("outs/repeat_with_no_schizophrenia/regions/astro_cn.pdf"), p1, width = 7, height = 4)

#FrCx 8
fc <- astro_de %>% filter(tissue == "FrCx") 
colour <-  mycoloursP[8]

p1 <- plotVolcano(fc, colour)
ggsave(here("outs/repeat_with_no_schizophrenia/regions/astro_frcx.pdf"), p1, width = 7, height = 4)
```


```{r eval=FALSE, include=FALSE}
# extract gene lists for venn
astro_de %>% filter(tissue == "FrCx") %>% .$gene %>% dput()
print("")
```

# Barplots
```{r}
oligo@meta.data %>% 
  mutate(tissue_status = paste(Tissue, donor_status, sep = "_")) %>% 
ggplot(aes(x = tissue_status, fill = ol_clusters_named_without_skf) ) + 
  geom_bar( position = "fill") + scale_fill_manual(values = mycoloursP) + coord_flip() + theme_classic() +
  labs(x = NULL,
       y = "Proportion",
       fill = NULL)
ggsave(here(paste0("outs/repeat_with_no_schizophrenia/abundance/prop_plot", "oligo.pdf")), width = 4, height = 3)

```

```{r}

astro@meta.data %>% 
  mutate(tissue_status = paste(Tissue, donor_status, sep = "_")) %>% 
ggplot(aes(x = tissue_status, fill = as_clusters_named_without_skf) ) + 
  geom_bar( position = "fill") + scale_fill_manual(values = mycoloursP) + coord_flip() + theme_classic() +
  labs(x = NULL,
       y = "Proportion",
       fill = NULL)
ggsave(here(paste0("outs/repeat_with_no_schizophrenia/abundance/prop_plot", "astro.pdf")), width = 4, height = 3)
```


```{r}
mglia@meta.data %>% 
  mutate(tissue_status = paste(Tissue, donor_status, sep = "_")) %>% 
ggplot(aes(x = tissue_status, fill = as_clusters_named_without_skf) ) + 
  geom_bar( position = "fill") + scale_fill_manual(values = mycoloursP) + coord_flip() + theme_classic() +
  labs(x = NULL,
       y = "Proportion",
       fill = NULL)
ggsave(here(paste0("outs/repeat_with_no_schizophrenia/abundance/prop_plot", "mglia.pdf")), width = 4, height = 3)

```
